+++ ek/platforms/android/java/ek/Sharing.java
@@ -14,14 +14,11 @@ public class Sharing {
     public static void navigate(final String url) {
         final EkActivity activity = EkActivity.getInstance();
         activity.runOnUiThread(
-                new Runnable() {
-                    @Override
-                    public void run() {
-                        try {
-                            Intent rateIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
-                            activity.startActivity(rateIntent);
-                        } catch (Exception ignored) {
-                        }
+                () -> {
+                    try {
+                        Intent rateIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
+                        activity.startActivity(rateIntent);
+                    } catch (Exception ignored) {
                     }
                 });
     }
@@ -30,18 +27,15 @@ public class Sharing {
     public static void rate_us(final String appId) {
         final EkActivity activity = EkActivity.getInstance();
         activity.runOnUiThread(
-                new Runnable() {
-                    @Override
-                    public void run() {
-                        try {
-                            final Bundle bundle = new Bundle();
-                            bundle.putString(FirebaseAnalytics.Param.ITEM_ID, "market");
-                            Analytics.logEvent("view_item", bundle);
+                () -> {
+                    try {
+                        final Bundle bundle = new Bundle();
+                        bundle.putString(FirebaseAnalytics.Param.ITEM_ID, "market");
+                        Analytics.logEvent("view_item", bundle);
 
-                            Intent rateIntent = new Intent(Intent.ACTION_VIEW, Uri.parse("market://details?id=" + appId));
-                            activity.startActivity(rateIntent);
-                        } catch (Exception ignored) {
-                        }
+                        Intent rateIntent = new Intent(Intent.ACTION_VIEW, Uri.parse("market://details?id=" + appId));
+                        activity.startActivity(rateIntent);
+                    } catch (Exception ignored) {
                     }
                 });
     }
@@ -50,22 +44,19 @@ public class Sharing {
     public static void send_message(final String text) {
         final EkActivity activity = EkActivity.getInstance();
         activity.runOnUiThread(
-                new Runnable() {
-                    @Override
-                    public void run() {
-                        try {
+                () -> {
+                    try {
 
-                            final Bundle bundle = new Bundle();
-                            bundle.putString(FirebaseAnalytics.Param.CONTENT_TYPE, "text");
-                            Analytics.logEvent("share", bundle);
+                        final Bundle bundle = new Bundle();
+                        bundle.putString(FirebaseAnalytics.Param.CONTENT_TYPE, "text");
+                        Analytics.logEvent("share", bundle);
 
-                            Intent intent = new Intent();
-                            intent.setAction(Intent.ACTION_SEND);
-                            intent.setType("text/plain");
-                            intent.putExtra(Intent.EXTRA_TEXT, text);
-                            activity.startActivity(Intent.createChooser(intent, "Share"));
-                        } catch (Exception ignored) {
-                        }
+                        Intent intent = new Intent();
+                        intent.setAction(Intent.ACTION_SEND);
+                        intent.setType("text/plain");
+                        intent.putExtra(Intent.EXTRA_TEXT, text);
+                        activity.startActivity(Intent.createChooser(intent, "Share"));
+                    } catch (Exception ignored) {
                     }
                 });
     }