+++ scenex/ek/goodies/bubble_text.cpp
@@ -15,7 +15,7 @@ void BubbleText::updateAll() {
     const float delta_y = -100.0f;
     const EaseOut<Back> back_out_5{Back{5.0f}};
 
-    for (auto e: ecs::rview<BubbleText>()) {
+    for (auto e: ecs::view_backward<BubbleText>()) {
         auto& state = ecs::get<BubbleText>(e);
 
         if (state.delay > 0.0f) {
@@ -40,10 +40,12 @@ void BubbleText::updateAll() {
             off = state.offset * easing::P3_OUT.calculate(sct);
         }
 
-        setScale(e, {sc, sc});
+        auto& transform = e.get<Transform2D>();
+        transform.scale = {sc, sc};
         float2 fly_pos{0.0f, delta_y * easing::P3_OUT.calculate(r)};
-        setPosition(e, state.start + off + fly_pos);
-        setAlpha(e, 1.0f - (r * r * r));
+        transform.position = state.start + off + fly_pos;
+        transform.color.setAlpha(1.0f - (r * r * r));
+        transform.color.setAdditive(r * r * r);
     }
 }
 
@@ -57,7 +59,7 @@ ecs::entity BubbleText::create(const float2& pos, const std::string& text, float
     TextFormat format{"Comfortaa-Regular", 32.0f};
     format.setAlignment(Alignment::Center);
     format.setTextColor(0xFFFFFFFF_argb);
-    format.addShadow(0x88000000_argb, 4, {0.0f, 1.5f});
+    format.addShadow(0xFF000000_argb, 4, {0.0f, 1.5f});
 
     e.assign<Display2D>(new Text2D(text, format));
     setTouchable(e, false);